move handler to user job request
