+++ server/controller/src/main/java/ai/starwhale/mlops/domain/job/JobCreator.java
@@ -23,27 +23,19 @@ import ai.starwhale.mlops.domain.job.bo.UserJobCreateRequest;
 import ai.starwhale.mlops.domain.job.cache.JobLoader;
 import ai.starwhale.mlops.domain.job.converter.UserJobConverter;
 import ai.starwhale.mlops.domain.job.po.JobFlattenEntity;
-import ai.starwhale.mlops.domain.job.spec.JobSpecParser;
-import ai.starwhale.mlops.domain.job.spec.StepSpec;
 import ai.starwhale.mlops.domain.job.split.JobSpliterator;
 import ai.starwhale.mlops.domain.job.status.JobStatus;
 import ai.starwhale.mlops.domain.job.status.JobUpdateHelper;
 import ai.starwhale.mlops.domain.storage.StoragePathCoordinator;
-import ai.starwhale.mlops.domain.system.SystemSettingService;
 import ai.starwhale.mlops.domain.upgrade.rollup.aspectcut.WriteOperation;
-import ai.starwhale.mlops.exception.SwValidationException;
-import ai.starwhale.mlops.exception.SwValidationException.ValidSubject;
-import ai.starwhale.mlops.exception.api.StarwhaleApiException;
+import ai.starwhale.mlops.exception.SwProcessException;
+import ai.starwhale.mlops.exception.SwProcessException.ErrorType;
 import cn.hutool.core.util.IdUtil;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import java.util.Date;
-import java.util.List;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.CollectionUtils;
-import org.springframework.util.StringUtils;
 
 
 @Component
@@ -55,8 +47,6 @@ public class JobCreator {
     private final JobDao jobDao;
     private final JobUpdateHelper jobUpdateHelper;
 
-    private final SystemSettingService systemSettingService;
-    private final JobSpecParser jobSpecParser;
     private final UserJobConverter userJobConverter;
 
     public JobCreator(
@@ -65,8 +55,6 @@ public class JobCreator {
             StoragePathCoordinator storagePathCoordinator,
             JobDao jobDao,
             JobUpdateHelper jobUpdateHelper,
-            SystemSettingService systemSettingService,
-            JobSpecParser jobSpecParser,
             UserJobConverter userJobConverter
     ) {
         this.jobSpliterator = jobSpliterator;
@@ -74,8 +62,6 @@ public class JobCreator {
         this.storagePathCoordinator = storagePathCoordinator;
         this.jobDao = jobDao;
         this.jobUpdateHelper = jobUpdateHelper;
-        this.systemSettingService = systemSettingService;
-        this.jobSpecParser = jobSpecParser;
         this.userJobConverter = userJobConverter;
     }
 
@@ -88,48 +74,15 @@ public class JobCreator {
         if (request instanceof UserJobCreateRequest) {
             builder = userJobConverter.convert((UserJobCreateRequest) request);
         } else {
-            builder = JobFlattenEntity.builder();
-        }
-
-        var stepSpecOverWrites = request.getStepSpecOverWrites();
-        var handler = request.getHandler();
-        if ((!StringUtils.hasText(stepSpecOverWrites) && !StringUtils.hasText(handler))
-                || (StringUtils.hasText(stepSpecOverWrites) && StringUtils.hasText(handler))) {
-            throw new StarwhaleApiException(
-                    new SwValidationException(ValidSubject.JOB, "handler or stepSpec must be provided only one"),
-                    HttpStatus.BAD_REQUEST
-            );
-        }
-        List<StepSpec> steps;
-        try {
-            steps = StringUtils.hasText(stepSpecOverWrites)
-                    ? jobSpecParser.parseAndFlattenStepFromYaml(stepSpecOverWrites)
-                    : jobSpecParser.parseStepFromYaml(builder.build().getModelVersion().getJobs(), handler);
-            for (var s : steps) {
-                s.verifyStepSpecArgs();
+            String stepSpec;
+            try {
+                stepSpec = Constants.yamlMapper.writeValueAsString(request.getStepSpecOverWrites());
+            } catch (JsonProcessingException e) {
+                throw new SwProcessException(ErrorType.SYSTEM, "serialize stepSpec failed", e);
             }
-            stepSpecOverWrites = Constants.yamlMapper.writeValueAsString(steps);
-        } catch (JsonProcessingException e) {
-            throw new StarwhaleApiException(
-                    new SwValidationException(ValidSubject.JOB, "failed to parse job step", e), HttpStatus.BAD_REQUEST);
-        }
-
-        if (CollectionUtils.isEmpty(steps)) {
-            throw new StarwhaleApiException(
-                    new SwValidationException(ValidSubject.JOB, "no stepSpec is configured"), HttpStatus.BAD_REQUEST);
-        }
-
-        var pool = systemSettingService.queryResourcePool(request.getResourcePool());
-        if (pool != null) {
-            for (var step : steps) {
-                pool.validateResources(step.getResources());
-            }
-            if (!pool.allowUser(request.getUser().getId())) {
-                throw new StarwhaleApiException(
-                        new SwValidationException(ValidSubject.JOB, "creator is not allowed to use this resource pool"),
-                        HttpStatus.BAD_REQUEST
-                );
-            }
+            builder = JobFlattenEntity.builder()
+                    .name(request.getStepSpecOverWrites().get(0).getJobName())
+                    .stepSpec(stepSpec);
         }
 
         JobFlattenEntity jobEntity = builder
@@ -137,14 +90,12 @@ public class JobCreator {
                 .project(request.getProject())
                 .projectId(request.getProject().getId())
                 .ownerId(request.getUser().getId())
-                .name(steps.get(0).getJobName())
                 .ownerName(request.getUser().getName())
                 .comment(request.getComment())
                 .resultOutputPath(storagePathCoordinator.allocateResultMetricsPath(jobUuid))
                 .jobStatus(JobStatus.CREATED)
                 .type(request.getJobType())
                 .resourcePool(request.getResourcePool())
-                .stepSpec(stepSpecOverWrites)
                 .createdTime(new Date())
                 .modifiedTime(new Date())
                 .build();
