+++ server/controller/src/main/java/ai/starwhale/mlops/domain/job/bo/UserJobCreateRequest.java
@@ -31,6 +31,8 @@ public class UserJobCreateRequest extends JobCreateRequest {
     @NotNull
     Long modelVersionId;
 
+    String handler;
+
     @NotNull
     Long runtimeVersionId;
