+++ server/controller/src/main/java/ai/starwhale/mlops/domain/job/converter/UserJobConverter.java
@@ -24,14 +24,20 @@ import ai.starwhale.mlops.domain.dataset.DatasetDao;
 import ai.starwhale.mlops.domain.dataset.bo.DatasetVersion;
 import ai.starwhale.mlops.domain.job.bo.UserJobCreateRequest;
 import ai.starwhale.mlops.domain.job.po.JobFlattenEntity;
+import ai.starwhale.mlops.domain.job.spec.JobSpecParser;
+import ai.starwhale.mlops.domain.job.spec.StepSpec;
 import ai.starwhale.mlops.domain.model.ModelDao;
 import ai.starwhale.mlops.domain.model.ModelService;
 import ai.starwhale.mlops.domain.project.ProjectService;
 import ai.starwhale.mlops.domain.runtime.RuntimeDao;
+import ai.starwhale.mlops.domain.system.SystemSettingService;
 import ai.starwhale.mlops.domain.user.UserService;
+import ai.starwhale.mlops.exception.SwProcessException;
+import ai.starwhale.mlops.exception.SwProcessException.ErrorType;
 import ai.starwhale.mlops.exception.SwValidationException;
 import ai.starwhale.mlops.exception.SwValidationException.ValidSubject;
 import ai.starwhale.mlops.exception.api.StarwhaleApiException;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
@@ -42,6 +48,7 @@ import javax.validation.constraints.NotNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 
 @Slf4j
@@ -57,6 +64,10 @@ public class UserJobConverter {
     private final RuntimeDao runtimeDao;
     private final DatasetDao datasetDao;
 
+    private final JobSpecParser jobSpecParser;
+
+    private final SystemSettingService systemSettingService;
+
     public UserJobConverter(
             IdConverter idConvertor,
             ProjectService projectService,
@@ -64,7 +75,9 @@ public class UserJobConverter {
             UserService userService,
             ModelDao modelDao,
             RuntimeDao runtimeDao,
-            DatasetDao datasetDao
+            DatasetDao datasetDao,
+            JobSpecParser jobSpecParser,
+            SystemSettingService systemSettingService
     ) {
         this.idConvertor = idConvertor;
         this.projectService = projectService;
@@ -73,6 +86,8 @@ public class UserJobConverter {
         this.modelDao = modelDao;
         this.runtimeDao = runtimeDao;
         this.datasetDao = datasetDao;
+        this.jobSpecParser = jobSpecParser;
+        this.systemSettingService = systemSettingService;
     }
 
     @NotNull
@@ -127,6 +142,12 @@ public class UserJobConverter {
             datasets = idConvertor.revertList(request.getDatasetVersionIds());
         }
 
+        List<StepSpec> stepSpecOverWrites = null;
+        try {
+            stepSpecOverWrites = jobSpecParser.parseAndFlattenStepFromYaml(request.getStepSpecOverWrites());
+        } catch (JsonProcessingException e) {
+            throw new SwValidationException(ValidSubject.MODEL, "invalid step spec", e);
+        }
         return UserJobCreateRequest.builder()
                 .project(project)
                 .modelVersionId(modelVersionId)
@@ -135,7 +156,7 @@ public class UserJobConverter {
                 .comment(request.getComment())
                 .resourcePool(request.getResourcePool())
                 .handler(request.getHandler())
-                .stepSpecOverWrites(request.getStepSpecOverWrites())
+                .stepSpecOverWrites(stepSpecOverWrites)
                 .bizType(request.getBizType())
                 .bizId(request.getBizId())
                 .jobType(request.getType())
@@ -163,6 +184,44 @@ public class UserJobConverter {
         }
         var runtime = runtimeDao.findById(runtimeVersion.getRuntimeId());
         var model = modelDao.findById(modelVersion.getModelId());
+        var stepSpecOverWrites = request.getStepSpecOverWrites();
+        var handler = request.getHandler();
+        if ((CollectionUtils.isEmpty(stepSpecOverWrites) && !StringUtils.hasText(handler))
+                || !(CollectionUtils.isEmpty(stepSpecOverWrites) && StringUtils.hasText(handler))) {
+            throw new StarwhaleApiException(
+                    new SwValidationException(ValidSubject.JOB, "handler or stepSpec must be provided only one"),
+                    HttpStatus.BAD_REQUEST
+            );
+        }
+        List<StepSpec> steps;
+        try {
+            steps = !CollectionUtils.isEmpty(stepSpecOverWrites)
+                    ? stepSpecOverWrites
+                    : jobSpecParser.parseStepFromYaml(modelVersion.getJobs(), handler);
+            for (var s : steps) {
+                s.verifyStepSpecArgs();
+            }
+        } catch (JsonProcessingException e) {
+            throw new StarwhaleApiException(
+                    new SwValidationException(ValidSubject.JOB, "failed to parse job step", e), HttpStatus.BAD_REQUEST);
+        }
+        if (CollectionUtils.isEmpty(steps)) {
+            throw new StarwhaleApiException(
+                    new SwValidationException(ValidSubject.JOB, "no stepSpec is configured"), HttpStatus.BAD_REQUEST);
+        }
+
+        var pool = systemSettingService.queryResourcePool(request.getResourcePool());
+        if (pool != null) {
+            for (var step : steps) {
+                pool.validateResources(step.getResources());
+            }
+            if (!pool.allowUser(request.getUser().getId())) {
+                throw new StarwhaleApiException(
+                        new SwValidationException(ValidSubject.JOB, "creator is not allowed to use this resource pool"),
+                        HttpStatus.BAD_REQUEST
+                );
+            }
+        }
 
         var datasetIds = request.getDatasetVersionIds();
         List<DatasetVersion> datasets = datasetIds != null ? datasetIds.stream()
@@ -190,6 +249,12 @@ public class UserJobConverter {
 
         var devMode = request.isDevMode();
 
+        String stepSpecYaml;
+        try {
+            stepSpecYaml = Constants.yamlMapper.writeValueAsString(steps);
+        } catch (JsonProcessingException e) {
+            throw new SwProcessException(ErrorType.SYSTEM, "serialize stepSpec failed", e);
+        }
         return JobFlattenEntity.builder()
                 .bizType(request.getBizType())
                 .bizId(request.getBizId())
@@ -206,6 +271,8 @@ public class UserJobConverter {
                 .runtimeVersionValue(runtimeVersion.getVersionName())
                 .modelName(model.getName())
                 .modelVersion(modelService.findModelVersion(modelVersion.getId()))
+                .stepSpec(stepSpecYaml)
+                .name(request.getStepSpecOverWrites().get(0).getJobName())
                 .modelVersionId(modelVersion.getId())
                 .modelVersionValue(modelVersion.getVersionName())
                 .modelUri(String.format(FORMATTER_URI_ARTIFACT, model.getProjectId(), "model", model.getId(),
